Fixed FxCop issues
